# Conflicts:
#	src/main/java/team/skyprojava/websitebackend/WebsiteBackendApplication.java
#	src/main/java/team/skyprojava/websitebackend/configuration/WebSecurityConfig.java
#	src/main/java/team/skyprojava/websitebackend/mapper/CommentMapper.java

+++ src/main/java/team/skyprojava/websitebackend/service/impl/AdsServiceImpl.java
@@ -4,7 +4,10 @@ package team.skyprojava.websitebackend.service.impl;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import team.skyprojava.websitebackend.dto.AdsDto;
@@ -12,19 +15,23 @@ import team.skyprojava.websitebackend.dto.CreateAdsDto;
 import team.skyprojava.websitebackend.dto.FullAdsDto;
 import team.skyprojava.websitebackend.dto.ResponseWrapperAdsDto;
 import team.skyprojava.websitebackend.entity.Ads;
+import team.skyprojava.websitebackend.entity.Comment;
 import team.skyprojava.websitebackend.entity.User;
 import team.skyprojava.websitebackend.exception.AdsNotFoundException;
 import team.skyprojava.websitebackend.exception.UserNotFoundException;
 import team.skyprojava.websitebackend.mapper.AdsMapper;
 import team.skyprojava.websitebackend.repository.AdsImageRepository;
 import team.skyprojava.websitebackend.repository.AdsRepository;
+import team.skyprojava.websitebackend.repository.CommentRepository;
 import team.skyprojava.websitebackend.repository.UserRepository;
+import team.skyprojava.websitebackend.security.SecurityAccess;
 import team.skyprojava.websitebackend.service.AdsService;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * Реализация сервиса для работы с объявлениями
@@ -37,6 +44,7 @@ public class AdsServiceImpl implements AdsService {
 
     private final Logger logger = LoggerFactory.getLogger(AdsServiceImpl.class);
     private final AdsRepository adsRepository;
+    private final CommentRepository commentRepository;
     private final AdsMapper adsMapper;
     private final UserRepository userRepository;
     private final AdsImageRepository adsImageRepository;
@@ -67,7 +75,11 @@ public class AdsServiceImpl implements AdsService {
     @Override
     public AdsDto createAds(CreateAdsDto createAdsDto, MultipartFile image, Authentication authentication) throws IOException {
         logger.info("Was invoked method for create ad");
-        User user = getUserByEmail(authentication.getName());
+        User user = getUserByEmail(SecurityContextHolder.getContext()
+                .getAuthentication().getName());
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new AccessDeniedException("Only authenticated users can create an ad");
+        }
         Ads ads = adsMapper.toEntity(createAdsDto);
         ads.setAuthor(user);
         //ads.setAdsImage(image);
@@ -76,18 +88,43 @@ public class AdsServiceImpl implements AdsService {
     }
 
     @Override
-    public void removeAds(int id) {
+    public boolean removeAds(int id, Authentication authentication) {
         logger.info("Was invoked method for delete ad by id");
-        Ads ads = getAdsById(id);
-        logger.warn("Ad by id {} not found", id);
-        adsRepository.delete(ads);
+        Ads ads = adsRepository.findById(id)
+                .orElseThrow(() -> new AdsNotFoundException("Объявление с id " + id + " не найдено!"));
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new AccessDeniedException("Only authenticated users can delete an ad");
+        }
+        if (!authentication.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))
+                && !ads.getAuthor().getEmail().equals(authentication.getName())) {
+            throw new AccessDeniedException("Only the author of the ad or an admin can delete it");
+        }
+        else if  (authentication.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))
+                && ads.getAuthor().getEmail().equals(authentication.getName())) {
+            List<Integer> adsComments = commentRepository.findAll().stream()
+                    .filter(adsComment -> adsComment.getAds().getId() == ads.getId())
+                    .map(Comment::getId)
+                    .collect(Collectors.toList());
+            commentRepository.deleteAllById(adsComments);
+            adsRepository.delete(ads);
+            logger.info("ad deleted");
+            return true;
+        }
+        return false;
     }
 
     @Override
-    public AdsDto updateAds(int id, CreateAdsDto updateAdsDto) {
+    public AdsDto updateAds(int id, CreateAdsDto updateAdsDto, Authentication authentication) {
         logger.info("Was invoked method for update ad by id");
+        // проверяем, что пользователь авторизован
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new AccessDeniedException("User is not authenticated");
+        }
         Ads updatedAds = getAdsById(id);
-        logger.warn("Ad by id {} not found", id);
+        // проверяем, что пользователь является автором объявления
+        if (!updatedAds.getAuthor().getEmail().equals(authentication.getName())) {
+            throw new AccessDeniedException("User is not the author of the ad");
+        }
         updatedAds.setTitle(updateAdsDto.getTitle());
         updatedAds.setDescription(updateAdsDto.getDescription());
         updatedAds.setPrice(updateAdsDto.getPrice());
@@ -97,8 +134,11 @@ public class AdsServiceImpl implements AdsService {
     }
 
     @Override
-    public FullAdsDto getFullAdsDto(int id) {
+    public FullAdsDto getFullAdsDto(int id, Authentication authentication) {
         logger.info("Was invoked method for get full ad dto");
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new AccessDeniedException("User is not authenticated");
+        }
         Ads ads = getAdsById(id);
         return adsMapper.toDto(ads);
     }
@@ -106,6 +146,9 @@ public class AdsServiceImpl implements AdsService {
     @Override
     public ResponseWrapperAdsDto getAdsMe(Authentication authentication) {
         logger.info("Service for get ads me");
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new AccessDeniedException("Don't authenticated");
+        }
         User user = getUserByEmail(authentication.getName());
         List<Ads> adsList = adsRepository.findAllByAuthorId(user.getId());
         if (!adsList.isEmpty()) {
@@ -118,7 +161,12 @@ public class AdsServiceImpl implements AdsService {
             result.setResults(adsDtoList);
             return result;
         }
-        throw new AdsNotFoundException("Ads are not found");
+        else {
+            ResponseWrapperAdsDto result = new ResponseWrapperAdsDto();
+            result.setCount(0);
+            result.setResults(Collections.emptyList());
+            return result;
+        }
     }
 
     @Override
